Moves portfolio card into new directory
